refactor: rename deprecated api
